Align rules and locking plugin interaction tests
